Move residency folder into service
