+++ src/dyad/CMakeLists.txt
@@ -7,4 +7,3 @@ add_subdirectory(client)
 add_subdirectory(service)
 add_subdirectory(wrapper)
 add_subdirectory(stream)
-add_subdirectory(residency)
